@@ -1,6 +1,6 @@
 name: flutter_bloc_cli
 description: CLI tool for rapidly scaffolding Flutter projects with BLoC architecture and code generation.
-version: 1.0.4
+version: 1.0.5
 repository: https://github.com/dharmeshgondaliya/flutter_bloc_cli
 
 platforms:
@@ -9,13 +9,13 @@ platforms:
   windows:
 
 environment:
-  sdk: ^3.3.0
+  sdk: ^3.6.0
 
 dependencies:
-  dart_style: ^2.3.6
-  dcli: ^6.0.5
-  process_run: ^0.14.2
-  pubspec: ^2.3.0
+  dart_style: ^3.0.1
+  dcli: ^6.1.0
+  process_run: ^1.2.2+1
+  pubspec_parse: ^1.4.0
 
 dev_dependencies:
   lints: ^3.0.0
